backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
mongomock-motor
openai
python-dotenv
apscheduler
